GG/AP - #210 - add View, Controller, Service and test for Controller and Service

+++ src/main/java/com/trailblazers/freewheelers/service/ItemService.java
@@ -25,5 +25,5 @@ public interface ItemService {
 
     ServiceResult<Item> saveItem(Item item);
 
-    int checkItemsQuantityIsMoreThanZero(long itemId);
+    long checkItemsQuantityIsMoreThanZero(long itemId);
 }
